feat: improved the mechanism for using luaunit_tts.lua. Now single point of entry.
